+++ index.tsx
@@ -5,10 +5,10 @@ import {AppRegistry} from 'react-native';
 import {name as appName} from './app.json';
 import {Provider} from 'react-redux';
 import store from './src/shared/presentation/redux/store';
-import AppWrapper from './src/shared/local_data/app_wrapper';
+import App from './src/App';
 
 const Index = () => {
-  return <Provider store={store} children={<AppWrapper />} />;
+  return <Provider store={store} children={<App />} />;
 };
 
 AppRegistry.registerComponent(appName, () => Index);
