+++ src/core/utils/constants/constants.ts
@@ -1,4 +1,4 @@
 // Constant object defining schema names used in a Realm database
 export const SCHEMA_NAMES = {
-  POST: 'Post', // Schema name for user-related data
+    POST: 'Post',
 };
