+++ src/features/posts/data/data_sources/post_data_source.ts
@@ -1,7 +1,6 @@
-import ApiResponse from '../../../../core/models/api_response'; // Importing ApiResponse model
-import PostDto from '../dto/post_dto'; // Importing PostDto class
+import ApiResponse from '../../../../core/models/api_response';
+import PostDto from '../dto/post_dto';
 
-// Interface for the data source responsible for retrieving post data API response from global core API response model
 export interface PostDataSource {
   getPosts(): Promise<ApiResponse<PostDto[]>>;
   getPost(): Promise<ApiResponse<PostDto>>;
