--- src/features/posts/domain/entities/post_entity.ts
@@ -1,9 +1,0 @@
-import {BSON} from 'realm';
-
-// IPost interface expecting to be used as per business logic on presentation layer
-export interface IPost {
-  userId: string;
-  _id: BSON.ObjectId;
-  title: string;
-  body: string;
-}
