--- src/shared/local_data/app_wrapper.tsx
@@ -1,7 +1,0 @@
-import RealmWrapper from './realm_wrapper';
-
-const AppWrapper = () => {
-  return <RealmWrapper />;
-};
-
-export default AppWrapper;
